collision manager broken out
